+++ lib/firmata.js
@@ -1,4 +1,5 @@
 import {Emitter} from "./events.js"
+import * as Types from "./types.js"
 
 // Polyfill
 import {Buffer} from "./buffer-es6/index.js"
@@ -428,46 +429,6 @@ const SYSEX_RESPONSE = {
 
 };
 
-export const MODES = {
-  INPUT: 0x00,
-  OUTPUT: 0x01,
-  ANALOG: 0x02,
-  PWM: 0x03,
-  SERVO: 0x04,
-  SHIFT: 0x05,
-  I2C: 0x06,
-  ONEWIRE: 0x07,
-  STEPPER: 0x08,
-  SERIAL: 0x0A,
-  PULLUP: 0x0B,
-  IGNORE: 0x7F,
-  PING_READ: 0x75,
-  UNKOWN: 0x10,
-}
-
-export const STEPPER_TYPES = {
-  TYPE: {
-    DRIVER: 1,
-    TWO_WIRE: 2,
-    THREE_WIRE: 3,
-    FOUR_WIRE: 4,
-  },
-  STEP_SIZE: {
-    WHOLE: 0,
-    HALF: 1
-  },
-  RUN_STATE: {
-    STOP: 0,
-    ACCEL: 1,
-    DECEL: 2,
-    RUN: 3,
-  },
-  DIRECTION: {
-    CCW: 0,
-    CW: 1,
-  },
-}
-
 /**
  * @class The Board object represents an arduino board.
  * @augments EventEmitter
@@ -506,7 +467,7 @@ export class Firmata extends Emitter {
 
     this.isReady = false;
 
-    this.MODES = MODES;
+    this.MODES = Types.MODES;
 
     this.I2C_MODES = {
       WRITE: 0,
@@ -515,7 +476,7 @@ export class Firmata extends Emitter {
       STOP_READING: 3,
     };
 
-    this.STEPPER = STEPPER_TYPES;
+    this.STEPPER = Types.STEPPER;
 
     this.SERIAL_MODES = {
       CONTINUOUS_READ: 0x00,
